Refactor out outlet mapping works-- to make the layers more explicit
Now with this refactor processing_node_types are the only ones that define inlets/outlets, and know what is allowed to connect to what. In addition they map specific return values to a given outlet. Then the processing_node_instance is responsible for mapping outlets to (node_id, inlet_id).